Initial SignalR Integration
Added all the components on the api server side to provide SignalR
functionality. Created a basic hub and implemented enough code identity
wise so  that users can get an idea on how SignalR can be used with
Taiseer's great project.

On the client side, I added a manually created hubs file for SignalR. I
also added JQuery and the SignalR required js files.

I created a simple service that provides functionality for integrating
with the SignalR server. It makes use of all the great stuff that
Taiseer has implemented.

I then for a simple demo added the service to the orders controller and
activate the SignalR service from there.

The server is configured to send a message back to the client once a
connection is made, for this example I simply when receiving those
messages from the server, I use Angulars broadcast service and listen
for those events in the orders controller.